new update 2017-05-04 update

+++ src/main/webapp/module/settings/settings.init.js
@@ -3,6 +3,7 @@ angular.module('settings_init', []).controller('settings_init', function($rootSc
     $scope.size = 15;
     $scope.pricetagtotal = 0;
     $scope.pricetags = [];
+    $scope.selectedWarehouse = {};
     $scope.find = function() {
         var fun = 'findAll';
         $http.get('pricetag/'+fun+'?page='+$scope.page+'&size='+$scope.size).then(function(response) {
@@ -36,11 +37,50 @@ angular.module('settings_init', []).controller('settings_init', function($rootSc
         $('#pricetag_modal').modal('show');
     };
     $scope.delete = function(item) {
-        $http.delete('pricetag/delete?id='+item.id).then(function(response) {
+        bootbox.confirm("Та үнийн төрлийн мэдээллийг устгах гэж байгаадаа итгэлтэй байна уу?",function(r) {
+            if (r) {
+                $http.delete('pricetag/delete?id=' + item.id).then(function (response) {
+                    $scope.find();
+                }, function (response) {
+                });
+            }
+        });
+    };
+    $scope.dialog_warehouse = function(item){
+        $scope.selectedWarehouse = item;
+        $('#warehouse_modal').modal('show');
+    };
+    $scope.add_warehouse = function(){
+        $scope.selectedWarehouse = {
+            id: 0,
+            name: '',
+            type:'storage',
+            owner:''
+        };
+        $("#warehouse_modal").modal('show');
+    };
+    $scope.save_warehouse = function(item){
+        $http.post('warehouse/save',item).then(function(response) {
             $scope.find();
         }, function(response) {
         });
     };
-
+    $scope.update_warehouse = function(item){
+        $http.put('warehouse/update',item).then(function(response) {
+            $("warehouse_modal").modal('hide');
+            $rootScope.getWarehouseList();
+        }, function(response) {
+        });
+    };
+    $scope.delete_warehouse = function(item){
+        bootbox.confirm("Та агуулахын мэдээллийг устгах гэж байгаадаа итгэлтэй байна уу?",function(r){
+            if(r){
+                $http.delete('warehouse/delete?id='+item.id).then(function(response) {
+                    $rootScope.getWarehouseList();
+                }, function(response) {
+                });
+            }
+        });
+    };
 //pr
 });
